+++ src/pages/Detail/Detail.tsx
@@ -1,270 +1,21 @@
 import styled from 'styled-components'
-import Slider from 'react-slick'
-import 'slick-carousel/slick/slick.css'
-import 'slick-carousel/slick/slick-theme.css'
-import Masonry from 'react-masonry-css'
-import BackDropImage from '../../assets/img/detail_backdrop.jpg'
-import PosterImage from '../../assets/img/detail_poster.jpg'
-import Actor01 from '../../assets/img/detail_actor01.jpg'
-import Actor02 from '../../assets/img/detail_actor02.jpg'
-import Actor03 from '../../assets/img/detail_actor03.jpg'
-import Actor04 from '../../assets/img/detail_actor04.jpg'
-import Actor05 from '../../assets/img/detail_actor05.jpg'
-import Actor06 from '../../assets/img/detail_actor06.jpg'
-import Actor07 from '../../assets/img/detail_actor07.jpg'
-import Actor08 from '../../assets/img/detail_actor08.jpg'
 
-import { useEffect, useRef, useState } from 'react'
 import PageLayout from '../../components/Layouts/PageLayout'
-import { Badge, BadgeBox, StyleAge, StyleTitle } from '../../components/Badge/style'
-import { formatLikes } from '../../utils/formatLikes'
-import { Checkbox, CheckboxWrapper, CheckHeartCount } from '../../components/Checkbox/style'
-import { AddIconBtn, MainBtn } from '../../components/Button/style'
-import { Textarea, TextareaBox } from '../../components/Input/style'
-import ImageUpload from '../../components/Input/ImageUpload/ImageUpload'
-import StarRating from '../../components/StarRating/StarRating'
-import useRating from '../../hooks/useRating'
-import ProfileBadge from '../../components/Badge/ProfileBadge/ProfileBadge'
-import { useInfiniteQuery, useQuery } from '@tanstack/react-query'
-import { client } from '../../apis/instances'
-import { useInView } from 'react-intersection-observer'
-
-interface Actor {
-  name: string
-  image: string
-}
-
-interface DummyData {
-  id: number
-  backdropImage: string
-  posterImage: string
-  title: string
-  age: number
-  genre: string[]
-  duration: number
-  playing: boolean
-  overview: string
-  trailer: string
-  actor: Actor[]
-  isLikes: boolean
-  totalLikes: number
-}
-
-const dummyData: DummyData = {
-  id: 0,
-  backdropImage: '이미지 경로',
-  posterImage: '이미지 경로',
-  title: '더 크로우',
-  age: 18,
-  genre: ['액션', 'SF/판타지'],
-  duration: 200,
-  playing: true,
-  overview:
-    '"더 크로우"는 에릭 드레븐이라는 뮤지션이 사랑하는 여자친구 셸리를 잔인하게 살해당한 후, 해골 같은 새의 힘으로 부활하는 이야기입니다. 에릭은 자신의 죽음을 기억하며 복수의 길에 나섭니다. 범죄자들을 처치하며 셸리와의 사랑을 되찾으려 하고, 자신의 과거와 마주합니다. 이 과정에서 그는 진정한 정의와 사랑의 의미를 찾으려 합니다. 에릭은 복수의 사냥꾼으로서, 그의 고통과 상실을 통해 인간의 복잡한 감정을 드러내며, 관객에게 깊은 여운을 남깁니다.',
-  trailer: '트레일러 경로',
-  actor: [
-    { name: '빌 스카스가드', image: Actor01 },
-    { name: 'FKA twigs', image: Actor02 },
-    { name: '대니 휴스턴', image: Actor03 },
-    { name: 'Laura Birn', image: Actor04 },
-    { name: 'Jordan Bolger', image: Actor05 },
-    { name: 'Isabella Wei', image: Actor06 },
-    { name: 'Karel Dobrý', image: Actor07 },
-    { name: 'Sebastian Orozco', image: Actor08 },
-  ],
-  isLikes: true,
-  totalLikes: 2500,
-}
-
-// interface Review {
-//   id: number
-//   userProfile: string
-//   userId: string
-//   score: number
-//   title: string
-//   content: string
-//   reviewImage: string
-//   creationDate: string
-// }
-
-const fetchReviewData = async (currentPage: any) => {
-  const res = await client.get(`/api/v1/review?page=${currentPage}`)
-  // const res = await client.get(`/api/v1/review?page=2}`)
-  const data = res.data
-  return data
-}
-
-const useGetPopularMovies = () => {
-  return useInfiniteQuery({
-    queryKey: ['popular-movie'],
-    queryFn: ({ pageParam }) => {
-      return fetchReviewData(pageParam)
-    },
-    getNextPageParam: (last) => {
-      if (last.currentPage < last.totalPages) {
-        return last.currentPage + 1
-      }
-      return undefined
-    },
-    initialPageParam: 1,
-  })
-}
+import { StyleTitle } from '../../components/Badge/style'
+import MovieDetailInfo from './ui/MovieDetailInfo'
+import ReviewWriteBox from './ui/ReviewWriteBox'
+import UserReviewList from './ui/UserReviewList'
 
 const Detail = () => {
-  const [movie, setMovie] = useState(dummyData)
-  const [image1, setImage1] = useState<string | null>(null)
-  const { rating: userRating1, handleRatingChange: handleRatingChange1 } = useRating()
-  const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } = useGetPopularMovies()
-
-  const { ref, inView } = useInView()
-
-  useEffect(() => {
-    if (inView && hasNextPage && !isFetchingNextPage) {
-      fetchNextPage()
-    }
-  }, [inView])
-
-  const settings = {
-    dots: true,
-    infinite: false,
-    speed: 500,
-    slidesToShow: 6,
-    slidesToScroll: 1,
-  }
-
   return (
     <DetailPage>
-      <button>버튼</button>
-      {/* 백드롭 */}
-      <BackDropImgBox>
-        <BackDropImg src={BackDropImage} alt={movie.title} />
-      </BackDropImgBox>
       <PageLayout>
-        {/* 영화 상세 */}
-        <MovieDetailWrapper>
-          <MovieDetails>
-            <PosterImgBox>
-              <PosterImg src={PosterImage} alt={movie.title} />
-            </PosterImgBox>
-            <MovieInfo>
-              <MovieInfoTop>
-                <MovieTitleHoder>
-                  <MovieTitle>
-                    <StyleAge $age='18' />
-                    {movie.title} 3
-                  </MovieTitle>
-                  <Badge $playing='playing'>상영중</Badge>
-                </MovieTitleHoder>
-                <BadgeBox>
-                  {movie.genre.map((item) => {
-                    return <Badge key={item}>{item}</Badge>
-                  })}
-                  <Badge $color='orange'>{movie.duration}분</Badge>
-                </BadgeBox>
-                <MovieOverviwe>{movie.overview}</MovieOverviwe>
-              </MovieInfoTop>
-              <MovieInfoBottom>
-                <MainBtn $size='large'>예매하기</MainBtn>
-                <CheckboxWrapper>
-                  <Checkbox type='checkbox' id='CheckHeartCount' name='' checked={movie.isLikes} />
-                  <CheckHeartCount htmlFor='CheckHeartCount' $color=''>
-                    {formatLikes(movie.totalLikes)}
-                  </CheckHeartCount>
-                </CheckboxWrapper>
-              </MovieInfoBottom>
-            </MovieInfo>
-          </MovieDetails>
-
-          {/* 트레일러 */}
-          <MovieTrailerBox>
-            <StyleTitle>Trailer</StyleTitle>
-            <MovieTrailer
-              width='100%'
-              style={{ aspectRatio: '16 / 9', height: 'auto' }}
-              src='https://www.youtube.com/embed/djSKp_pwmOA?si=l1wO5dvCTL_p6bKx'
-              title='YouTube video player'
-              frameBorder='0'
-              allow='accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share'
-              referrerPolicy='strict-origin-when-cross-origin'
-              allowFullScreen
-            ></MovieTrailer>
-          </MovieTrailerBox>
-
-          {/* 배우 */}
-          <MovieActorBox>
-            <StyleTitle>Actor</StyleTitle>
-            <MovieActorList>
-              <Slider {...settings}>
-                {movie.actor.map((actor) => (
-                  <MovieActor key={actor.name}>
-                    <MovieActorImg
-                      src={actor.image}
-                      alt={actor.name}
-                      style={{ width: '100%', height: 'auto' }}
-                    />
-                    <MovieActorName>{actor.name}</MovieActorName>
-                  </MovieActor>
-                ))}
-              </Slider>
-            </MovieActorList>
-          </MovieActorBox>
-        </MovieDetailWrapper>
-
-        {/* 리뷰 */}
+        <MovieDetailInfo />
         <MovieReviewWrapper>
           <StyleTitle>Review</StyleTitle>
-          <MovieReviewForm>
-            <TextareaBox width=''>
-              <Textarea width='' placeholder='텍스트를 입력하세요.' />
-            </TextareaBox>
-
-            <ReviewInputWrapper>
-              <StarRating
-                rating={userRating1}
-                onRatingChange={handleRatingChange1}
-                userId='user1'
-              />
-              <ActionBtnBox>
-                <ImageUpload image={image1} setImage={setImage1} />
-                <AddIconBtn $size='large'>작성하기</AddIconBtn>
-              </ActionBtnBox>
-            </ReviewInputWrapper>
-          </MovieReviewForm>
-
-          <UserRiviewWrapper>
-            <Masonry
-              breakpointCols={4} // 화면 크기에 따라 열 수 조정
-              className='my-masonry-grid' // 클래스 이름
-              columnClassName='my-masonry-grid_column' // 각 열의 클래스 이름
-            >
-              {data?.pages.map((page) =>
-                page.reviews.map((review) => (
-                  <UserRiviewBox key={review?.id}>
-                    <RiviewImgBox>
-                      <RiviewImg src={review?.reviewImage} alt='사람들이 업로드한 이미지' />
-                    </RiviewImgBox>
-                    <RiviewInfoBox>
-                      <RiviewUserInfo>
-                        <UserInfoHolder>
-                          <ProfileBadge src='' width='3.2rem' height='3.2rem' />
-
-                          <UserInfoBox>
-                            <UserId>{review.userId}</UserId>
-                            <CreationDate>{review.creationDate}</CreationDate>
-                          </UserInfoBox>
-                        </UserInfoHolder>
-                        <StarRating value={review.score} disabled={true} />
-                      </RiviewUserInfo>
-                      <RiviewTitle>{review.title}</RiviewTitle>
-                      <RiviewContent>{review.content}</RiviewContent>
-                    </RiviewInfoBox>
-                  </UserRiviewBox>
-                )),
-              )}
-            </Masonry>
-          </UserRiviewWrapper>
+          <ReviewWriteBox />
+          <UserReviewList />
         </MovieReviewWrapper>
-        <LoadReview ref={ref} />
       </PageLayout>
     </DetailPage>
   )
@@ -274,229 +25,7 @@ export default Detail
 
 const DetailPage = styled.div``
 
-const BackDropImgBox = styled.span`
-  display: block;
-  margin: 0 auto;
-`
-const BackDropImg = styled.img`
-  width: 100%;
-`
-
-const MovieDetailWrapper = styled.div``
-
-// 영화 상세
-const MovieDetails = styled.div`
-  display: flex;
-  gap: 6rem;
-  padding-top: 6rem;
-`
-const PosterImgBox = styled.span`
-  display: block;
-  width: 100%;
-  max-width: 40rem;
-  margin-top: -14rem;
-`
-const PosterImg = styled.img`
-  width: 100%;
-`
-const MovieInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-`
-const MovieInfoTop = styled.div``
-const MovieTitleHoder = styled.div`
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  padding-bottom: 2rem;
-  margin-bottom: 1.2rem;
-  border-bottom: 1px solid #3f3f3f;
-`
-const MovieTitle = styled.p`
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-  font-size: 3.2rem;
-  font-weight: 700;
-`
-const MovieOverviwe = styled.p`
-  margin-top: 2rem;
-  font-size: 2rem;
-  font-weight: 500;
-  line-height: 1.5;
-  word-break: keep-all;
-`
-const MovieInfoBottom = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-`
-
-// 트레일러
-const MovieTrailerBox = styled.div`
-  margin-top: 10rem;
-`
-const MovieTrailer = styled.iframe`
-  margin-top: 2.4rem;
-`
-
-// 배우
-const MovieActorBox = styled.div`
-  margin-top: 10rem;
-`
-const MovieActorList = styled.div`
-  margin-top: 2.4rem;
-  div:focus-visible {
-    outline: none;
-  }
-  .slick-track {
-    display: flex;
-    gap: 1.8rem;
-    width: 100%;
-
-    &::before,
-    &::after {
-      display: none;
-    }
-  }
-  .slick-arrow {
-    display: none !important;
-    position: absolute;
-  }
-  .slick-dots {
-    li {
-      width: auto;
-      height: auto;
-      margin: 0 4px;
-      button {
-        content: '';
-        position: initial;
-        display: block;
-        width: 8px;
-        height: 8px;
-        padding: 0;
-        line-height: initial;
-        background-color: #353535;
-        border: 1px solid;
-        border-color: #353535;
-        border-radius: 4px;
-        opacity: 1;
-        transition: all 0.3s;
-        &::before {
-          display: none;
-        }
-      }
-
-      &.slick-active {
-        button {
-          width: 12px;
-          background-color: #f45917;
-          border-color: #ff8b59;
-          transition: all 0.3s;
-        }
-      }
-    }
-  }
-`
-const MovieActor = styled.div``
-const MovieActorImg = styled.img``
-const MovieActorName = styled.p`
-  margin-top: 1.2rem;
-  font-size: 1.8rem;
-  font-weight: 600;
-`
-
 // 리뷰
 const MovieReviewWrapper = styled.div`
   margin-top: 10rem;
-`
-const MovieReviewForm = styled.div`
-  padding-bottom: 2rem;
-  margin-top: 2.4rem;
-  margin-bottom: 4rem;
-  border-bottom: 1px solid #3f3f3f;
-`
-const ReviewInputWrapper = styled.div`
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin-top: 1.2rem;
-`
-const ActionBtnBox = styled.div`
-  display: flex;
-  align-items: flex-start;
-  gap: 0.8rem;
-`
-
-const UserRiviewWrapper = styled.div`
-  .my-masonry-grid {
-    display: flex;
-    margin-left: -24px; /* 그리드 간격 조정 */
-    width: auto; /* 자동 너비 */
-  }
-
-  .my-masonry-grid_column {
-    padding-left: 24px; /* 각 열의 간격 */
-    background-clip: padding-box; /* 배경 클립 */
-  }
-`
-const UserRiviewBox = styled.div`
-  margin-bottom: 2.4rem;
-  background-color: #1b1b1b;
-  border: 1px solid #494949;
-  border-radius: 0.8rem;
-  overflow: hidden;
-  transition: transform 0.3s;
-
-  &:hover {
-    transform: scale(1.1);
-    transition: transform 0.3s;
-  }
-`
-
-const RiviewImgBox = styled.span``
-const RiviewImg = styled.img`
-  width: 100%;
-`
-
-const RiviewInfoBox = styled.div`
-  padding: 1.2rem;
-`
-const RiviewUserInfo = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`
-
-const UserInfoHolder = styled.span`
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-`
-const UserInfoBox = styled.span`
-  display: flex;
-  flex-direction: column;
-  gap: 0.4rem;
-`
-const UserId = styled.span`
-  font-size: 1.6rem;
-`
-const CreationDate = styled.span`
-  font-size: 1.2rem;
-`
-
-const RiviewTitle = styled.p`
-  margin-top: 1.2rem;
-  font-size: 1.8rem;
-  font-weight: 700;
-`
-const RiviewContent = styled.p`
-  margin-top: 0.8rem;
-  font-size: 1.6rem;
-  line-height: 1.25;
-`
-const LoadReview = styled.div`
-  height: 8rem;
 `
