--- src/pages/Detail/ui/UI.tsx
@@ -1,1 +1,0 @@
-// UI
